requirements.txt

streamlit==1.26.0
crewai==0.28.0
langchain-openai==0.0.5
langchain==0.1.0
langchain-community==0.0.13
openai==1.12.0
duckduckgo-search==3.9.5
newspaper3k==0.2.8
pandas==2.1.1
matplotlib==3.8.0
plotly==5.18.0
python-dotenv==1.0.0
faiss-cpu==1.7.4